Slettet Datainnsamling_Idunn og endret navn på luftkvalitet
